Added genre, editor, author, manytomany relation book genre

+++ src/main/java/com/tolrom/springlibrary/exception/BookNotFoundException.java
@@ -1,7 +1,7 @@
 package com.tolrom.springlibrary.exception;
 
 public class BookNotFoundException extends RuntimeException {
-    public BookNotFoundException(Long id) {
+    public BookNotFoundException(Integer id) {
         super("The book with id : " + id + " doesn't exist");
     }
 }
